refactor(dashboard): enhance filter functionality by consolidating filter-related types and helpers across dashlets
