+++ apps/app/src/features/dashboard/dashlets/data_table/dashlet.settings.tsx
@@ -17,16 +17,17 @@ import type {
   DashletConfig,
   TableColumn,
   ColumnType,
-  FilterConfig,
-  FilterItemConfig,
   SortConfig,
 } from "./dashlet";
 import {
   defaultColumns,
   defaultRows,
   defaultSort,
-  normalizeFilterConfig,
+  defaultFilter,
 } from "./dashlet";
+import type { FilterConfig, FilterItemConfig } from "@/features/dashboard/dashlets/common/filter-types";
+import { normalizeFilterConfig, toFilterItems, fromFilterItems } from "@/features/dashboard/dashlets/common/filter-helpers";
+import type { FilterItem } from "@/features/dashboard/dashlets/common/filter-helpers";
 import { SettingsTextField, SettingsSelectField } from "../common";
 import type { ColumnItem } from "@/features/dashboard/dashlets/common/column-helpers";
 import { toColumnItems, fromColumnItems } from "@/features/dashboard/dashlets/common/column-helpers";
@@ -38,22 +39,6 @@ import { tr } from "@/features/i18n/tr.service";
 // ============================================================================
 
 type SettingsTab = "visualization" | "data";
-
-interface FilterItem extends FilterItemConfig {
-  _id: string;
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function toFilterItems(items: FilterItemConfig[]): FilterItem[] {
-  return items.map((item, i) => ({ ...item, _id: `fi-${i}-${item.column}` }));
-}
-
-function fromFilterItems(items: FilterItem[]): FilterItemConfig[] {
-  return items.map(({ column, label }) => ({ column, label }));
-}
 
 // ============================================================================
 // Component
@@ -81,7 +66,7 @@ export function DashletSettings({
   );
 
   // Filter config (normalize legacy shapes)
-  const normalizedFilter = normalizeFilterConfig(config.filter);
+  const normalizedFilter = normalizeFilterConfig(config.filter, defaultFilter);
   const [filterEnabled, setFilterEnabled] = useState(normalizedFilter.enabled);
   const [filterItems, setFilterItems] = useState<FilterItem[]>(
     toFilterItems(normalizedFilter.items)
