+++ apps/app/src/features/dashboard/dashlets/data_table/dashlet.tsx
@@ -1,32 +1,25 @@
 "use client";
 
-import { useState, useMemo } from "react";
+import { useMemo } from "react";
 import { HiArrowUp, HiArrowDown } from "react-icons/hi2";
 import type { DashletComponentProps, DashletLayoutDefaults } from "@/features/dashboard/dashlets/types";
 import type { TableColumn, SortConfig } from "@/features/dashboard/dashlets/common/column-types";
+import type { FilterConfig, FilterItemConfig } from "@/features/dashboard/dashlets/common/filter-types";
 import { renderCell } from "@/features/dashboard/dashlets/common/cell-renderers";
 import { Pill } from "@/features/dashboard/dashlets/common/pill";
 import { useDynamicRows } from "@/features/dashboard/dashlets/common/use-dynamic-rows";
+import { normalizeFilterConfig } from "@/features/dashboard/dashlets/common/filter-helpers";
+import { FilterPillRow } from "@/features/dashboard/dashlets/common/filter-pill-row";
+import { useFilterAndSort } from "@/features/dashboard/dashlets/common/use-filter-and-sort";
 import { useDashboard } from "@/features/dashboard/context/dashboard-context";
 import { tr } from "@/features/i18n/tr.service";
 
 export type { ColumnType, TableColumn, SortConfig } from "@/features/dashboard/dashlets/common/column-types";
+export type { FilterItemConfig, FilterConfig } from "@/features/dashboard/dashlets/common/filter-types";
 
 // ============================================================================
 // Configuration Types
 // ============================================================================
-
-export interface FilterItemConfig {
-  /** Column key whose distinct values are used as filter pills */
-  column: string;
-  /** Label shown before the filter pills, e.g. "Estado:" */
-  label: string;
-}
-
-export interface FilterConfig {
-  enabled: boolean;
-  items: FilterItemConfig[];
-}
 
 export interface DashletConfig {
   title: string;
@@ -97,37 +90,7 @@ export const defaultFilter: FilterConfig = {
   items: [{ column: "status", label: "Estado:" }],
 };
 
-/**
- * Normalize a persisted filter config into the current shape.
- * Old configs stored `{ enabled, column, label }` — convert to `{ enabled, items }`.
- */
-export function normalizeFilterConfig(raw: unknown): FilterConfig {
-  if (!raw || typeof raw !== "object") return defaultFilter;
-  const obj = raw as Record<string, unknown>;
-  const enabled = typeof obj.enabled === "boolean" ? obj.enabled : defaultFilter.enabled;
-  if (Array.isArray(obj.items)) {
-    const validItems = obj.items.filter(
-      (item): item is FilterItemConfig =>
-        !!item &&
-        typeof item === "object" &&
-        typeof (item as Record<string, unknown>).column === "string" &&
-        (item as Record<string, unknown>).column !== ""
-    ).map((item) => ({
-      column: item.column,
-      label: typeof item.label === "string" ? item.label : "",
-    }));
-    if (validItems.length === 0) return defaultFilter;
-    return { enabled, items: validItems };
-  }
-  // Legacy shape: { enabled, column, label }
-  if (typeof obj.column === "string" && obj.column !== "") {
-    return {
-      enabled,
-      items: [{ column: obj.column, label: typeof obj.label === "string" ? obj.label : "" }],
-    };
-  }
-  return defaultFilter;
-}
+export { normalizeFilterConfig } from "@/features/dashboard/dashlets/common/filter-helpers";
 
 export const defaultSort: SortConfig = {
   enabled: true,
@@ -159,49 +122,6 @@ export function getLayoutDefaults(): DashletLayoutDefaults {
 }
 
 // ============================================================================
-// Filter Pill Row
-// ============================================================================
-
-interface FilterPillRowProps {
-  item: FilterItemConfig;
-  options: string[];
-  selected: string;
-  allLabel: string;
-  onClear: (column: string) => void;
-  onSelect: (column: string, value: string) => void;
-}
-
-function FilterPillRow({
-  item,
-  options,
-  selected,
-  allLabel,
-  onClear,
-  onSelect,
-}: Readonly<FilterPillRowProps>) {
-  return (
-    <div className="flex shrink-0 flex-wrap items-center gap-2 rounded-lg border border-gray-200 bg-white px-4 py-3 dark:border-gray-700 dark:bg-gray-800">
-      <span className="text-sm text-gray-500 dark:text-gray-400">
-        {item.label}
-      </span>
-      <Pill
-        label={allLabel}
-        active={selected === ""}
-        onClick={() => onClear(item.column)}
-      />
-      {options.map((val) => (
-        <Pill
-          key={val}
-          label={val}
-          active={selected === val}
-          onClick={() => onSelect(item.column, val)}
-        />
-      ))}
-    </div>
-  );
-}
-
-// ============================================================================
 // Component
 // ============================================================================
 
@@ -217,83 +137,28 @@ export function Dashlet({ widget }: Readonly<DashletComponentProps>) {
     apiUrl = "",
     sort = defaultSort,
   } = config;
-  const filter = useMemo(() => normalizeFilterConfig(config.filter), [config.filter]);
+  const filter = useMemo(() => normalizeFilterConfig(config.filter, defaultFilter), [config.filter]);
 
   // ── Dynamic data fetching ───────────────────────────────────────────────────
   const { rows: dynamicRows, loading, fetchError } = useDynamicRows(dataMode, apiUrl);
 
-  // ── Filter & sort state ─────────────────────────────────────────────────────
-  const [filterValues, setFilterValues] = useState<Record<string, string>>({});
-  const [sortKey, setSortKey] = useState<string | null>(null);
-  const [sortDir, setSortDir] = useState<"asc" | "desc">("asc");
-
   const allRows = dataMode === "dynamic" ? dynamicRows : staticRows;
 
-  // Distinct values per filter item (derived from full dataset)
-  const filterOptionsByColumn = useMemo(() => {
-    if (!filter.enabled) return {};
-    const result: Record<string, string[]> = {};
-    for (const item of filter.items) {
-      const seen = new Set<string>();
-      for (const row of allRows) {
-        const val = row[item.column];
-        if (val) seen.add(val);
-      }
-      result[item.column] = Array.from(seen);
-    }
-    return result;
-  }, [allRows, filter.enabled, filter.items]);
-
-  // Column label lookup for sort toolbar
-  const getColumnLabel = (key: string) =>
-    columns.find((c) => c.key === key)?.label ?? key;
-
-  // Apply all active filters (AND) then sort
-  const displayRows = useMemo(() => {
-    let result = allRows;
-
-    if (filter.enabled) {
-      for (const item of filter.items) {
-        const selected = filterValues[item.column];
-        if (selected) {
-          result = result.filter((row) => row[item.column] === selected);
-        }
-      }
-    }
-
-    if (sort.enabled && sortKey) {
-      result = [...result].sort((a, b) => {
-        const cmp = (a[sortKey] ?? "").localeCompare(b[sortKey] ?? "");
-        return sortDir === "asc" ? cmp : -cmp;
-      });
-    }
-
-    return result;
-  }, [allRows, filter, filterValues, sort, sortKey, sortDir]);
+  // ── Filter & sort (shared hook) ───────────────────────────────────────────
+  const {
+    filterValues,
+    sortKey,
+    sortDir,
+    filterOptionsByColumn,
+    displayRows,
+    getColumnLabel,
+    handleFilterClear,
+    handleFilterSelect,
+    handleSortClick,
+  } = useFilterAndSort(filter, sort, allRows, columns);
 
   const getSortIcon = (dir: "asc" | "desc") =>
     dir === "asc" ? <HiArrowUp className="h-3 w-3" /> : <HiArrowDown className="h-3 w-3" />;
-
-  const handleFilterClear = (column: string) => {
-    setFilterValues((prev) => {
-      const next = { ...prev };
-      delete next[column];
-      return next;
-    });
-  };
-
-  const handleFilterSelect = (column: string, value: string) => {
-    setFilterValues((prev) => ({ ...prev, [column]: value }));
-  };
-
-  const handleSortClick = (key: string) => {
-    if (sortKey === key) {
-      setSortDir((d) => (d === "asc" ? "desc" : "asc"));
-    } else {
-      setSortKey(key);
-      setSortDir("asc");
-    }
-  };
 
   // ── Render ──────────────────────────────────────────────────────────────────
   const allLabel = tr("common.all", dictionary);
@@ -315,7 +180,7 @@ export function Dashlet({ widget }: Readonly<DashletComponentProps>) {
 
       {/* Filter cards */}
       {filter.enabled &&
-        filter.items.map((item, idx) => {
+        filter.items.map((item: FilterItemConfig, idx: number) => {
           const options = filterOptionsByColumn[item.column];
           if (!options || options.length === 0) return null;
           return (
